edit character show character sheet
